Migrate `gene-annotations-track.test.js`
